Move results to a new app, migrate models and fix API
